[Vim] Pressing 'esc' in the CLI should focus the editor. Closes #487
Pressing ESC with the Vim mode activated takes you now to the editor,
in good Vim fashion. It conflicts a bit with the previous behavior,
which was that pressing ESC would hide a hint in case there was one
displayed. Now it will hide the hint AND take you to the editor.