[Vim] Paste doesn't work when deleting in visual mode. Closes #490
Deleting in visual mode now yanks properly.

Other small fixes to the way paste works are in this commits to make it
more accurate to the 'vim' way. The `Util` module is now holding the
visual mode state, which is not ideal but works nicely, since that one
is the only module without any dependency.