[Vim] Cursor out of text. Closes #495
This fixes the weird behavior on empty lines and line endings. Vim
works in a very different way than ace, and besides having different
column position indexing (starting with 1), doesn't necessarily jump to
the next/previous line on selection or moving right/left.

So, what this patch tries to do is to detect where the cursor is and
tweak the behavior. It might not yet be perfect (right now it behaves
nicely but the `w` command doesn't jump to the next line, which should
be filed as another bug).