Restore default keybindings taken away by Vim. Closes #568
This patch fixes the problem related to the overwriting
of commands like `CMD-/` after the Vim plugin has been 
enabled.

As a side effect of this patch, some performance 
enhancements have happened, given that we are not
listening to the (buggy) `pref.checked` event anymore.
This event was causing listeners to be fired very
frequently while disabling Vim mode every time, even
when it was already disabled.

Furthermore, Vim mode is not listening to the
unnecessary "afteropenfile" event anymore.

Some other minor fixes have gone in in files related
to this patch, where globals and other issues pointed
out by the language module have been solved.

A new copy of Ace is referenced because it solves 
an important bug related to the `getKeyboardHandler`
method.