Make margin preservation async in vim mode
DRYed and ensured synchronousness when re-scrolling, since Ace's search is asynchronous and cancels the scrolling in some cases. Margin preservation is clearly something that Ace should do at some point. Probably passing a property to the `options` object passed to the `findPrev`/`findNext` methods. Left as an exercise for the reader.