Changed the method by which nodes are added:
Instead of adding them to the tree, they are added
to the model. This is because FF interpreted the "add"
operation to mean we were trying to create a new folder
or file, and borked. Luckily adding the nodes
to the model does not affect our ability to expand
the tree nodes and we still get the great sliding animation
as the nodes are being populated.

In addition fixed the last-file-selected issue.