Improving fix #1 - Launching a Terminal removes the console command bar
Tested with chrome and safari.

Closing the console from the X button (right side) re-displays the command bar (but not when disabled from View->Console). When re-opening console from the lower right corner button (next to command line) will re-hide the command line. If command was written to command line the console is opened in Console tab and the command line remains shown.