Additional fix #1 - Stops switching to terminal tab if console is hidden
If console is closed in terminal tab and a refresh is  performed, the console remains hidden but hidden switch to terminal page was done, disabling the command line in the process. This stops the terminal tab switch in case the console itself is hidden for the user session setting. To continue from here, the user needs to enable console and manually click he terminal tab to see the restored session terminals (if any).